+++ docker/dockernetwork/dockernetwork_mock.go
@@ -1,0 +1,26 @@
+package dockernetwork
+
+import (
+	"context"
+	"github.com/docker/engine-api/client"
+	"github.com/docker/engine-api/types"
+	"github.com/docker/engine-api/types/network"
+	"github.com/stretchr/testify/mock"
+)
+
+type MockApiClient struct {
+	mock.Mock
+	client.Client
+}
+
+func (m *MockApiClient) NetworkList(ctx context.Context, options types.NetworkListOptions) ([]types.NetworkResource, error) {
+	args := m.Called(ctx, options)
+	return args.Get(0).([]types.NetworkResource), args.Error(1)
+}
+
+func (*MockApiClient) CreateMockNetwork(ip string, id string) *network.EndpointSettings {
+	m := new(network.EndpointSettings)
+	m.IPAddress = ip
+	m.NetworkID = id
+	return m
+}
